fix controlnet with fp16 again

+++ api/onnx_web/diffusers/pipelines/controlnet.py
@@ -416,9 +416,9 @@ class OnnxStableDiffusionControlNetPipeline(DiffusionPipeline):
                 timestep = np.array([t], dtype=timestep_dtype)
 
                 blocksamples = self.controlnet(
-                    sample=latent_model_input.astype(np.float32),
+                    sample=latent_model_input,
                     timestep=timestep,
-                    encoder_hidden_states=prompt_embeds.astype(np.float32),
+                    encoder_hidden_states=prompt_embeds,
                     controlnet_cond=image,
                 )
